Remove Chat page and rename CopilotKit page to Chat
- Delete original Chat page at /chat
- Move CopilotKit page from /copilot to /chat
- Rename CopilotPage to ChatPage and update header title
- Remove CopilotKit entry from sidebar navigation
- Remove unused Sparkles icon import
