+++ apps/ui/src/app/chat/page.tsx
@@ -1,124 +1,194 @@
 "use client";
 
-import { useState, useCallback, useEffect } from "react";
+import { useState, useCallback, useEffect, useRef } from "react";
 import { useAgents } from "@/hooks/use-agents";
-import { useCreateThread, useMessages, useCreateMessage } from "@/hooks/use-threads";
-import { useCreateRun, useRun } from "@/hooks/use-runs";
-import { useSSEEvents, aggregateTextMessages, aggregateToolCalls } from "@/hooks/use-sse-events";
 import { Header } from "@/components/layout/header";
 import { AgentSelector } from "@/components/chat/agent-selector";
-import { ChatMessages } from "@/components/chat/chat-messages";
-import { ChatInput } from "@/components/chat/chat-input";
 import { Card, CardContent } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import { Skeleton } from "@/components/ui/skeleton";
-import { Bot, Plus } from "lucide-react";
+import { Bot, Plus, Send, User } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import { Textarea } from "@/components/ui/textarea";
+import { ScrollArea } from "@/components/ui/scroll-area";
 import Link from "next/link";
+
+const API_BASE = process.env.NEXT_PUBLIC_API_BASE_URL || "http://localhost:9000";
+
+interface ChatMessage {
+  id: string;
+  role: "user" | "assistant";
+  content: string;
+}
+
+// AG-UI event types
+interface AgUiEvent {
+  type: string;
+  threadId?: string;
+  runId?: string;
+  messageId?: string;
+  delta?: string;
+  message?: string;
+}
 
 export default function ChatPage() {
   // Agent selection
   const { data: agents = [], isLoading: agentsLoading } = useAgents();
   const [selectedAgentId, setSelectedAgentId] = useState<string | null>(null);
 
-  // Thread state
+  // Chat state
+  const [messages, setMessages] = useState<ChatMessage[]>([]);
+  const [input, setInput] = useState("");
+  const [isStreaming, setIsStreaming] = useState(false);
+  const [streamingContent, setStreamingContent] = useState("");
   const [threadId, setThreadId] = useState<string | null>(null);
-  const { data: messages = [] } = useMessages(threadId || "");
-  const createThread = useCreateThread();
-  const createMessage = useCreateMessage(threadId || "");
+  const scrollRef = useRef<HTMLDivElement>(null);
 
-  // Run state
-  const [currentRunId, setCurrentRunId] = useState<string | null>(null);
-  const { data: currentRun } = useRun(currentRunId || "");
-  const createRun = useCreateRun();
-
-  // SSE events for current run
-  const { events } = useSSEEvents({
-    runId: currentRunId || "",
-    enabled: !!currentRunId && currentRun?.status === "running",
-  });
-
-  // Derived state
-  const streamingMessages = aggregateTextMessages(events);
-  const streamingToolCalls = aggregateToolCalls(events);
-  const isStreaming = !!currentRunId && currentRun?.status === "running";
-  const isWaiting = createThread.isPending || createMessage.isPending || createRun.isPending;
-
-  // Clear current run when it finishes
+  // Auto-scroll to bottom
   useEffect(() => {
-    if (currentRun && (currentRun.status === "completed" || currentRun.status === "failed")) {
-      // Keep the run ID for a moment to show final state, then clear
-      const timer = setTimeout(() => {
-        setCurrentRunId(null);
-      }, 500);
-      return () => clearTimeout(timer);
+    if (scrollRef.current) {
+      scrollRef.current.scrollTop = scrollRef.current.scrollHeight;
     }
-  }, [currentRun?.status]);
+  }, [messages, streamingContent]);
 
   // Handle agent selection
   const handleAgentChange = useCallback((agentId: string) => {
     setSelectedAgentId(agentId);
-    // Reset conversation when agent changes
+    setMessages([]);
     setThreadId(null);
-    setCurrentRunId(null);
   }, []);
 
   // Handle sending a message
-  const handleSendMessage = useCallback(
-    async (content: string) => {
-      if (!selectedAgentId) return;
+  const handleSend = useCallback(async () => {
+    if (!input.trim() || !selectedAgentId || isStreaming) return;
 
-      try {
-        // Create thread if needed
-        let tid = threadId;
-        if (!tid) {
-          const thread = await createThread.mutateAsync();
-          tid = thread.id;
-          setThreadId(tid);
-        }
+    const userMessage: ChatMessage = {
+      id: crypto.randomUUID(),
+      role: "user",
+      content: input.trim(),
+    };
 
-        // Add user message
-        await createMessage.mutateAsync({
-          role: "user",
-          content,
-        });
+    setMessages((prev) => [...prev, userMessage]);
+    setInput("");
+    setIsStreaming(true);
+    setStreamingContent("");
 
-        // Create run
-        const run = await createRun.mutateAsync({
-          agent_id: selectedAgentId,
-          thread_id: tid,
-        });
-
-        setCurrentRunId(run.id);
-      } catch (error) {
-        console.error("Failed to send message:", error);
+    try {
+      // Build the AG-UI endpoint URL
+      const params = new URLSearchParams({
+        agent_id: selectedAgentId,
+      });
+      if (threadId) {
+        params.set("thread_id", threadId);
       }
-    },
-    [selectedAgentId, threadId, createThread, createMessage, createRun]
-  );
 
-  // Handle new conversation
-  const handleNewConversation = useCallback(() => {
-    setThreadId(null);
-    setCurrentRunId(null);
-  }, []);
+      const url = `${API_BASE}/v1/ag-ui?${params}`;
+
+      // Send request with messages
+      const response = await fetch(url, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          messages: messages.concat(userMessage).map((m) => ({
+            role: m.role,
+            content: m.content,
+          })),
+        }),
+      });
+
+      if (!response.ok) {
+        throw new Error(`HTTP error: ${response.status}`);
+      }
+
+      // Get thread ID from response headers
+      const responseThreadId = response.headers.get("X-Thread-Id");
+      if (responseThreadId) {
+        setThreadId(responseThreadId);
+      }
+
+      // Parse SSE stream
+      const reader = response.body?.getReader();
+      if (!reader) throw new Error("No response body");
+
+      const decoder = new TextDecoder();
+      let buffer = "";
+      let currentContent = "";
+      let currentMessageId = "";
+
+      while (true) {
+        const { done, value } = await reader.read();
+        if (done) break;
+
+        buffer += decoder.decode(value, { stream: true });
+
+        // Process complete SSE events
+        const lines = buffer.split("\n");
+        buffer = lines.pop() || ""; // Keep incomplete line in buffer
+
+        for (const line of lines) {
+          if (line.startsWith("data: ")) {
+            const jsonStr = line.slice(6);
+            try {
+              const event: AgUiEvent = JSON.parse(jsonStr);
+              console.log("AG-UI Event:", event);
+
+              switch (event.type) {
+                case "RunStarted":
+                  if (event.threadId) {
+                    setThreadId(event.threadId);
+                  }
+                  break;
+
+                case "TextMessageStart":
+                  currentMessageId = event.messageId || crypto.randomUUID();
+                  currentContent = "";
+                  break;
+
+                case "TextMessageContent":
+                  if (event.delta) {
+                    currentContent += event.delta;
+                    setStreamingContent(currentContent);
+                  }
+                  break;
+
+                case "TextMessageEnd":
+                  if (currentContent) {
+                    const assistantMessage: ChatMessage = {
+                      id: currentMessageId,
+                      role: "assistant",
+                      content: currentContent,
+                    };
+                    setMessages((prev) => [...prev, assistantMessage]);
+                    setStreamingContent("");
+                  }
+                  break;
+
+                case "RunFinished":
+                case "RunError":
+                  setIsStreaming(false);
+                  break;
+              }
+            } catch {
+              // Ignore parse errors for keepalive comments
+            }
+          }
+        }
+      }
+
+      setIsStreaming(false);
+    } catch (error) {
+      console.error("Failed to send message:", error);
+      setIsStreaming(false);
+    }
+  }, [input, selectedAgentId, messages, threadId, isStreaming]);
 
   const selectedAgent = agents.find((a) => a.id === selectedAgentId);
   const canChat = !!selectedAgentId;
 
   return (
     <>
-      <Header
-        title="Chat"
-        action={
-          threadId && (
-            <Button variant="outline" onClick={handleNewConversation}>
-              <Plus className="h-4 w-4 mr-2" />
-              New Conversation
-            </Button>
-          )
-        }
-      />
+      <Header title="Chat" />
       <div className="flex flex-col h-[calc(100vh-64px)]">
         {agentsLoading ? (
           <div className="p-6">
@@ -164,56 +234,92 @@ export default function ChatPage() {
                 </div>
               </div>
             ) : (
-              <>
-                {/* Thread Info Bar */}
-                {threadId && (
-                  <div className="px-4 py-2 bg-muted/50 border-b flex items-center gap-2 text-sm">
-                    <span className="text-muted-foreground">Thread:</span>
-                    <Link
-                      href={`/threads/${threadId}`}
-                      className="font-mono text-primary hover:underline"
-                    >
-                      {threadId.slice(0, 8)}...
-                    </Link>
-                    {currentRunId && (
-                      <>
-                        <span className="text-muted-foreground">|</span>
-                        <span className="text-muted-foreground">Run:</span>
-                        <Link
-                          href={`/runs/${currentRunId}`}
-                          className="font-mono text-primary hover:underline"
-                        >
-                          {currentRunId.slice(0, 8)}...
-                        </Link>
-                        {isStreaming && (
-                          <Badge variant="outline" className="bg-blue-100 text-blue-800 animate-pulse">
-                            Streaming
-                          </Badge>
+              <div className="flex-1 flex flex-col">
+                {/* Messages */}
+                <ScrollArea className="flex-1 p-4" ref={scrollRef}>
+                  <div className="space-y-4 max-w-3xl mx-auto">
+                    {messages.map((message) => (
+                      <div
+                        key={message.id}
+                        className={`flex gap-3 ${
+                          message.role === "user" ? "justify-end" : "justify-start"
+                        }`}
+                      >
+                        {message.role === "assistant" && (
+                          <div className="w-8 h-8 rounded-full bg-primary/10 flex items-center justify-center">
+                            <Bot className="h-4 w-4 text-primary" />
+                          </div>
                         )}
-                      </>
+                        <div
+                          className={`rounded-lg px-4 py-2 max-w-[80%] ${
+                            message.role === "user"
+                              ? "bg-primary text-primary-foreground"
+                              : "bg-muted"
+                          }`}
+                        >
+                          <p className="whitespace-pre-wrap">{message.content}</p>
+                        </div>
+                        {message.role === "user" && (
+                          <div className="w-8 h-8 rounded-full bg-primary flex items-center justify-center">
+                            <User className="h-4 w-4 text-primary-foreground" />
+                          </div>
+                        )}
+                      </div>
+                    ))}
+
+                    {/* Streaming message */}
+                    {streamingContent && (
+                      <div className="flex gap-3 justify-start">
+                        <div className="w-8 h-8 rounded-full bg-primary/10 flex items-center justify-center">
+                          <Bot className="h-4 w-4 text-primary animate-pulse" />
+                        </div>
+                        <div className="rounded-lg px-4 py-2 max-w-[80%] bg-muted">
+                          <p className="whitespace-pre-wrap">{streamingContent}</p>
+                        </div>
+                      </div>
+                    )}
+
+                    {/* Loading indicator */}
+                    {isStreaming && !streamingContent && (
+                      <div className="flex gap-3 justify-start">
+                        <div className="w-8 h-8 rounded-full bg-primary/10 flex items-center justify-center">
+                          <Bot className="h-4 w-4 text-primary animate-pulse" />
+                        </div>
+                        <div className="rounded-lg px-4 py-2 bg-muted">
+                          <p className="text-muted-foreground">Thinking...</p>
+                        </div>
+                      </div>
                     )}
                   </div>
-                )}
-
-                {/* Messages */}
-                <ChatMessages
-                  messages={messages}
-                  streamingMessages={streamingMessages}
-                  streamingToolCalls={streamingToolCalls}
-                  isStreaming={isStreaming}
-                />
+                </ScrollArea>
 
                 {/* Input */}
-                <ChatInput
-                  onSend={handleSendMessage}
-                  disabled={isStreaming || isWaiting}
-                  placeholder={
-                    isStreaming
-                      ? "Waiting for response..."
-                      : `Message ${selectedAgent?.name || "agent"}...`
-                  }
-                />
-              </>
+                <div className="border-t p-4">
+                  <div className="max-w-3xl mx-auto flex gap-2">
+                    <Textarea
+                      value={input}
+                      onChange={(e) => setInput(e.target.value)}
+                      placeholder={`Message ${selectedAgent?.name || "agent"}...`}
+                      disabled={isStreaming}
+                      className="min-h-[44px] max-h-32 resize-none"
+                      onKeyDown={(e) => {
+                        if (e.key === "Enter" && !e.shiftKey) {
+                          e.preventDefault();
+                          handleSend();
+                        }
+                      }}
+                    />
+                    <Button
+                      onClick={handleSend}
+                      disabled={!input.trim() || isStreaming}
+                      size="icon"
+                      className="h-11 w-11"
+                    >
+                      <Send className="h-4 w-4" />
+                    </Button>
+                  </div>
+                </div>
+              </div>
             )}
           </>
         )}
